Move `appendPlatformTypeTestName` and `appendDriverVestionTestName` in `utils.hpp` instead of `npu_test_env_cfg.hpp`
