+++ src/plugins/intel_npu/tests/functional/common/npu_test_env_cfg.hpp
@@ -56,17 +56,6 @@ std::string getDeviceNameTestCase(const std::string& str);
 std::string getDeviceName();
 std::string getDeviceNameID(const std::string& str);
 
-// Current version of gtest seems to fail when parsing template functions for getTestCaseName
-// To overcome this issue, programmer must make sure to wrap in paratheses this function when
-// it is given at INSTANTIATE_TEST_SUITE_P macros
-// e.g. INSTANTIATE_TEST_SUITE_P(TEST_SUITE_NAME, TEST_SUITE_CLASS, params,
-// (appendPlatformTypeTestName<TEST_SUITE_CLASS>))
-template <typename T>
-std::string appendPlatformTypeTestName(testing::TestParamInfo<typename T::ParamType> obj) {
-    const std::string& test_name = GenericTestCaseNameClass::getTestCaseName<T>(obj);
-    return test_name + "_targetPlatform=" + getTestsPlatformFromEnvironmentOr(ov::test::utils::DEVICE_NPU);
-}
-
 }  // namespace ov::test::utils
 
 namespace InferRequestParamsAnyMapTestName {
