+++ src/plugins/intel_npu/tests/functional/shared_tests_instances/behavior/ov_plugin/core_threading_tests.cpp
@@ -5,7 +5,6 @@
 #include <utility>
 
 #include "behavior/ov_plugin/core_threading.hpp"
-#include "common/npu_test_env_cfg.hpp"
 #include "common/utils.hpp"
 #include "intel_npu/npu_private_properties.hpp"
 
@@ -23,35 +22,28 @@ const Params params_disable_umd_cache[] = {std::tuple<Device, Config>{
 
 const Params params_cached[] = {std::tuple<Device, Config>{ov::test::utils::DEVICE_NPU, {}}};
 
-template <typename T>
-std::string appendDriverVestionTestName(testing::TestParamInfo<typename T::ParamType> obj) {
-    const auto& pluginCacheCore = ov::test::utils::PluginCache::get().core(ov::test::utils::DEVICE_NPU);
-    auto driverVersion =
-        pluginCacheCore->get_property(ov::test::utils::DEVICE_NPU, ov::intel_npu::driver_version.name());
-    return ov::test::utils::appendPlatformTypeTestName<T>(obj) + "_driverVersion=" + driverVersion.as<std::string>();
-}
-
 }  // namespace
 
-INSTANTIATE_TEST_SUITE_P(compatibility_smoke_BehaviorTests_CoreThreadingTest_NPU,
-                         CoreThreadingTest,
-                         testing::ValuesIn(params),
-                         (appendDriverVestionTestName<CoreThreadingTest>));  // need to get also driver version to skip
-                                                                             // failing tests with PV driver
+INSTANTIATE_TEST_SUITE_P(
+    compatibility_smoke_BehaviorTests_CoreThreadingTest_NPU,
+    CoreThreadingTest,
+    testing::ValuesIn(params),
+    ov::test::utils::appendDriverVersionTestName<CoreThreadingTest>);  // need to get also driver version to skip
+                                                                       // failing tests with PV driver
 
 INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests_CoreThreadingTest_NPU,
                          CoreThreadingTestsWithIter,
                          testing::Combine(testing::ValuesIn(params), testing::Values(15), testing::Values(50)),
-                         (ov::test::utils::appendPlatformTypeTestName<CoreThreadingTestsWithIter>));
+                         ov::test::utils::appendPlatformTypeTestName<CoreThreadingTestsWithIter>);
 
 INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests_CoreThreadingTest_NPU,
                          CoreThreadingTestsWithCacheEnabled,
                          testing::Combine(testing::ValuesIn(params_cached), testing::Values(10), testing::Values(30)),
-                         (ov::test::utils::appendPlatformTypeTestName<CoreThreadingTestsWithCacheEnabled>));
+                         ov::test::utils::appendPlatformTypeTestName<CoreThreadingTestsWithCacheEnabled>);
 
 INSTANTIATE_TEST_SUITE_P(smoke_BehaviorTests_CoreThreadingTest_UmdCacheDisabled_NPU,
                          CoreThreadingTestsWithIter,
                          testing::Combine(testing::ValuesIn(params_disable_umd_cache),
                                           testing::Values(8),
                                           testing::Values(20)),
-                         (ov::test::utils::appendPlatformTypeTestName<CoreThreadingTestsWithIter>));
+                         ov::test::utils::appendPlatformTypeTestName<CoreThreadingTestsWithIter>);
