+++ src/plugins/intel_npu/tests/functional/shared_tests_instances/skip_tests.xml
@@ -677,6 +677,7 @@ Example:
         <filters>
             <filter>.*CoreThreadingTestsWithIter.smoke_CompileModel/.*</filter>
             <filter>.*CoreThreadingTestsWithIter.smoke_CompileModel_MultipleCores/.*</filter>
+            <filter>.*CoreThreadingTestsWithIter.nightly_AsyncInfer_ShareInput/.*</filter>
         </filters>
     </skip_config>
 
@@ -685,7 +686,7 @@ Example:
         <message>Failing Windows PV Driver core threading tests</message>
         <enable_rules>
             <backend>LEVEL0</backend>
-            <operating_system>linux</operating_system>
+            <operating_system>windows</operating_system>
         </enable_rules>
         <filters>
             <filter>.*CoreThreadingTest.smoke_QueryModel/.*_driverVersion=1688.*</filter>
